requirements.txt
Flask==3.0.0
opencv-python-headless==4.9.0.80
openai==1.12.0
python-dotenv==1.0.0
Pillow==10.1.0
werkzeug==3.0.1
gunicorn==21.2.0
azure-identity==1.15.0
numpy==1.24.3
httpx==0.24.1